QA-14864 : fix encoding
